fix(metrics): fix many bugs with metrics
- Loading the metrics from export
- New timm version (KLD)

+++ metrics/rhythm_similarity.py
@@ -68,4 +68,4 @@ def calculate_rhythm_similarity(
     if verbose:
         print(f"Rhythm similarity: {total_rhythm_similarity / count}")
 
-    return total_rhythm_similarity / count
+    return float(total_rhythm_similarity / count)
